requirements.txt
streamlit>=1.29.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
groq>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0